Phase3/requirements.txt
browser-use
anthropic
python-dotenv
playwright
fastapi
uvicorn
pydantic
pytest
pytest-asyncio
aiofiles